Connected sfCoursesChart
