+++ src/views/DashboardSchwundberechnung.js
@@ -22,7 +22,7 @@ function DashboardSchwundberechnung(props) {
               </Card.Body>
             </Card>
         </Col>
-        {/* <Col xs='5'>
+        <Col xs='5'>
             <Card className="card-chart">
               <Card.Body>
                 <div>
@@ -30,7 +30,7 @@ function DashboardSchwundberechnung(props) {
                 </div>
               </Card.Body>
             </Card>
-        </Col> */}
+        </Col>
         </Row>
         {/* <Row>
           <Col xs="10">
